==> notebooks/01-get-data-from-spotify.ipynb <==
# Buscar dados do Spotify

Esse notebook tem o intuito de buscar músicas do Spotify através de playlists de usuários para montar um dataset de recomendações.

### Bibliotecas

- spotipy
- pandas

### Instalação e importação de dependências

pip install spotipy python-dotenv


from spotipy.oauth2 import SpotifyClientCredentials
from dotenv import load_dotenv
import pandas as pd
import spotipy
import os

load_dotenv()

client_id = os.getenv("SPOTIFY_CLIENT_ID")
client_secret = os.getenv("SPOTIFY_CLIENT_SECRET")

client_credentials_manager = SpotifyClientCredentials(client_id, client_secret)
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)


Podemos fazer essa busca de duas formas:

1. Buscar as playlists com base em categorias do Spotify
2. Buscar as playlists de um usuário previamente autenticado

### Buscando com base nas categorias

categories = []  # Inserir as categorias
playlists = []

for category in categories:
    offset = 0
    limit = 20

    while True:
        page_results = sp.category_playlists(
            category, offset=offset, limit=limit)
        playlists += page_results["playlists"]["items"]

        if not page_results["playlists"]["next"]:
            break

        offset += limit


### Buscando com base no usuário

playlists = []
offset = 0
limit = 50

while True:
    page_results = sp.current_user_playlists(offset=offset, limit=limit)
    playlists += page_results["items"]

    if not page_results["next"]:
        break

    offset += limit


Agora que recebemos as playlists, podemos buscar as músicas delas.

### Criar uma lista com os IDs das playlists

playlists_ids = []
for playlist in playlists:
    if playlist is not None:
        playlists_ids.append(playlist["id"])

playlists_ids


### Buscar as músicas das playlists

fields = "items(track(name, artists(name, id, popularity), id, popularity)), next"
results = []

for playlist_id in playlists_ids:
    offset = 0
    limit = 50

    while True:
        page_results = sp.playlist_tracks(
            playlist_id, fields=fields, offset=offset, limit=limit)
        results += page_results["items"]

        if not page_results["next"]:
            break

        offset += limit

print(len(results))


tracks = []

for item in results:
    if item["track"] is not None:
        track = {
            "artist_name": str(item["track"]["artists"][0]["name"]),
            "artist_url": item["track"]["artists"][0]["id"],
            "track_name": item["track"]["name"],
            "track_url": item["track"]["id"],
            "track_popularity": item["track"]["popularity"]
        }
        tracks.append(track)
    else:
        print("Skipping item with missing track")

len(tracks)


df = pd.DataFrame(tracks)
df.head(10)


Agora com as informações das músicas, vamos salvar em um `.csv` para fazermos o pré-processamento e o processamento.

# Para um novo to_process.csv
df.to_csv("to_process.csv", index=False)
# Para mesclar em um to_process.csv já existente
# df.to_csv("to_process.csv", mode="a", index=False, header=False)
