==> notebooks/02-get-features-data.ipynb <==
# Pré-processar os dados do Spotify e extrair as características

Esse notebook tem o intuito de fazer o pré-processamento dos dados e também extrair as `features` para criação da nossa base de recomendações.

### Bibliotecas

- tqdm
- spotipy
- python-dotenv
- pandas
- multiprocessing

### Instalação e importação de dependências

pip install tqdm spotipy python-dotenv


import os
import pandas as pd
import spotipy
import multiprocessing
import concurrent.futures
from tqdm import tqdm
from spotipy.oauth2 import SpotifyClientCredentials
from dotenv import load_dotenv

load_dotenv()


client_id = os.getenv("SPOTIFY_CLIENT_ID")
client_secret = os.getenv("SPOTIFY_CLIENT_SECRET")

client_credentials_manager = SpotifyClientCredentials(client_id, client_secret)
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)


### Importação dos dados

Aqui utilizarems o arquivo `to_process.csv` gerado no notebook `01-get-data-from-spotify.ipynb` para fazer o pré-processamento e extração das características.

file_path = "../data/to_process.csv"
fields = ["artist_name", "track_name",
          "track_url", "artist_url", "track_popularity"]

df = pd.read_csv(file_path, usecols=fields)
df.head()


### Buscando e incluindo as `features`

Nesta etapa é montada as funções que utilizaremos para extrais as características das músicas através da API do Spotify.

def merge_track_artists_ids(df):
    df["track_artist_id"] = df.apply(lambda row: str(
        row["track_url"]) + "_" + str(row["artist_url"]), axis=1)
    return df


def process_tracks(ids):
    tracks_ids = []
    artists_ids = []

    for id in ids:
        [track_id, artist_id] = id.split("_")
        tracks_ids.append(track_id)
        artists_ids.append(artist_id)

    try:
        # Buscar as features
        audio_features = sp.audio_features(tracks_ids)

        # Buscar dados do artista
        artists = []
        for i in range(0, len(artists_ids), 50):
            artist_batch = artists_ids[i:i+50]
            artists += sp.artists(artist_batch)["artists"]

        tracks = []
        for i in range(len(tracks_ids)):
            track = audio_features[i]
            track.update({
                "artist_popularity": artists[i]["popularity"],
                "artist_genres": artists[i]["genres"]
            })
            tracks.append(track)

        return tracks
    except:
        return []


df = merge_track_artists_ids(df)
process_tracks([df["track_artist_id"][0]])


Para buscar as características iremos utilizar o `ThreadPoolExecutor` para paralelizar  as chamadas de função, visto que em alguns momentos podemos ter um fluxo alto de dados para processar.

Para saber a quantidade máxima de workers que podemos usar, contamos quantas threads temos disponíveis usando o `multiprocessing`.

cpu_count = multiprocessing.cpu_count()

max_workers = cpu_count if cpu_count < 12 else 12
[cpu_count, max_workers]


def chunks(data, chunk_size):
    for i in range(0, len(data), chunk_size):
        yield data[i:i + chunk_size]


unique_ids = df["track_artist_id"].unique()

# Usando o ThreadPoolExecutioner para processar as músicas paralelamente
with concurrent.futures.ThreadPoolExecutor(max_workers=max_workers) as executor:
    # Criar um executor com a funcão process_tracks para cada chunk de músicas
    future_results = [executor.submit(process_tracks, chunk)
                      for chunk in chunks(unique_ids, 100)]

    # Usar o tqdm para acompanhar o progresso dos executores
    features = []
    for future in tqdm(concurrent.futures.as_completed(future_results), total=len(future_results)):
        # Pegar os resultados e mesclar eles
        feature = future.result()
        features.extend(feature)


features_df = pd.DataFrame(features)
features_df.head()

### Finalização e exportação

Agora mesclamos os dataframes, excluimos as duplicatas e salvamos em um `.csv`

df.drop(columns=["track_artist_id"], inplace=True)

completed_df = pd.merge(df, features_df, left_on="track_url", right_on="id")
completed_df.head()

def drop_duplicates(df):
    df["artists_song"] = df.apply(lambda row: str(
        row["artist_name"]) + str(row["track_name"]), axis=1)
    df = df.drop_duplicates("artists_song")
    print("Are all songs unique:", len(pd.unique(df.artists_song)) == len(df))
    df = df.drop(columns=["artists_song"], inplace=True)
    return df


completed_df = drop_duplicates(completed_df)


# Para um novo processed_data.csv
df.to_csv("processed_data.csv", index=False)
# Para mesclar em um processed_data.csv já existente
# df.to_csv("processed_data.csv", mode="a", index=False, header=False)
