==> notebooks/03-process-data.ipynb <==
# Criar base de recomendação

Neste notebook iremos pegar os dados pré-processados e criar nossa base de recomendação.

### Bibliotecas

- pandas
- scikit-learn
- textblob

### Instalação e importação de dependências

pip install textblob


import pandas as pd
from sklearn.preprocessing import MinMaxScaler
from sklearn.feature_extraction.text import TfidfVectorizer
from textblob import TextBlob


### Buscando e organizando nossos dados

Nesta etapa buscamos nosso `.csv` criado no notebook `02-get-features-data.ipynb` e organizamos ele para criar nossa base de recomendação.

data_df = pd.read_csv("../data/processed_data.csv")
data_df.head()


def drop_duplicates(df):
    df['artists_song'] = df.apply(lambda row: str(
        row['artist_name']) + str(row['track_name']), axis=1)
    return df.drop_duplicates('artists_song')


def drop_duplicates_ids(df):
    return df.drop_duplicates("id")


tracks_df = drop_duplicates(data_df)
print("Are all songs unique:", len(
    pd.unique(tracks_df.artists_song)) == len(tracks_df))

tracks_df = drop_duplicates_ids(tracks_df)
print("Are all songs unique:", len(pd.unique(tracks_df.id)) == len(tracks_df))


Agora iremos filtrar apenas as colunas que utilizaremos para o sistema de recomendação, além disso vamos separar os dados em 3 categorias (`Metadata`, `Audio` e `Text`):

1. Metadata
    - id
    - genres
    - artist_popularity
    - track_popularity
2. Audio
    - **Mood:** danceability, valence, energy, tempo
    - **Properties:** loudness, speechiness, instrumentalness
    - **Context:** liveness, acousticness
    - **metadata:** key, mode
3. Text
    - track_name

def select_columns(df):
    return df[["artist_name", "id", "track_name", "danceability", "energy", "key", "loudness", "mode",
               "speechiness", "acousticness", "instrumentalness", "liveness", "valence", "tempo", "artist_popularity", "genres", "track_popularity"]]


tracks_df = select_columns(tracks_df)
tracks_df.head()


Com as colunas selecionadas, vamos começar organizando a coluna `genres`, que é a unica da categoria `Metadata` que precisa ser sanitizada.

def genre_preprocess(df):
    df["genres_list"] = df["genres"].apply(lambda x: str(x).split(" "))
    return df


tracks_df = genre_preprocess(data_df)
tracks_df["genres_list"].head()


Com os dados da categoria `Metadata` organizados, vamos trabalhar no único da categoria `Text`, que é o `track_name`

def get_subjectivity(text):
    return TextBlob(str(text)).sentiment.subjectivity


def get_polarity(text):
    return TextBlob(str(text)).sentiment.polarity


def get_subjectivity_analysis(score):
    if score < 1/3:
        return "low"
    elif score > 1/3:
        return "high"
    else:
        return "medium"


def get_polarity_analysis(score):
    if score < 0:
        return 'Negative'
    elif score == 0:
        return 'Neutral'
    else:
        return 'Positive'


def sentiment_analysis(df, column_name):
    df["subjectivity"] = df[column_name].apply(
        get_subjectivity).apply(lambda x: get_subjectivity_analysis(x))
    df["polarity"] = df[column_name].apply(
        get_polarity).apply(lambda x: get_polarity_analysis(x))

    return df


sentiment_df = sentiment_analysis(tracks_df, "track_name")
sentiment_df.head()

Agora partimos para a última categoria, a `Audio`

def ohe_preparation(df, column, new_name):
    term_frequency_df = pd.get_dummies(df[column])
    feature_names = term_frequency_df.columns

    term_frequency_df.columns = [
        new_name + "|" + str(i) for i in feature_names]
    term_frequency_df.reset_index(drop=True, inplace=True)

    return term_frequency_df


subject_ohe = ohe_preparation(sentiment_df, 'subjectivity', 'subject')
subject_ohe.iloc[0]


tf_idf = TfidfVectorizer()
tf_idf_matrix = tf_idf.fit_transform(
    tracks_df["genres_list"].apply(lambda x: " ".join(x)))

genre_df = pd.DataFrame(tf_idf_matrix.toarray())
genre_df.columns = ["genre" + "|" + i for i in tf_idf.get_feature_names_out()]
genre_df.drop(columns="genre|unknown")
genre_df.reset_index(drop=True, inplace=True)

genre_df.iloc[0]


print(tracks_df["artist_popularity"].describe())


popularity = tracks_df[["artist_popularity"]].reset_index(drop=True)
scaler = MinMaxScaler()

popularity_scaled = pd.DataFrame(scaler.fit_transform(
    popularity), columns=popularity.columns)
popularity_scaled.head()


Agora com todas as funções feitas, vamos criar o dataset de recomendações

def create_features_dataset(df, float_cols):
    tf_idf = TfidfVectorizer()
    tf_idf_matrix = tf_idf.fit_transform(
        df["genres_list"].apply(lambda x: " ".join(x)))

    genre_df = pd.DataFrame(tf_idf_matrix.toarray())
    genre_df.columns = ["genre" + "|" +
                        i for i in tf_idf.get_feature_names_out()]
    genre_df.drop(columns="genre|unknown")
    genre_df.reset_index(drop=True, inplace=True)

    df = sentiment_analysis(df, "track_name")

    subject_ohe = ohe_preparation(df, "subjectivity", "subject") * 0.3
    polar_ohe = ohe_preparation(df, "polarity", "polar") * 0.5
    key_ohe = ohe_preparation(df, "key", "key") * 0.5
    mode_ohe = ohe_preparation(df, "mode", "mode") * 0.5

    popularity = df[["artist_popularity", "track_popularity"]
                    ].reset_index(drop=True)
    scaler = MinMaxScaler()
    popularity_scaled = pd.DataFrame(scaler.fit_transform(
        popularity), columns=popularity.columns) * 0.2

    floats = df[float_cols].reset_index(drop=True)
    scaler = MinMaxScaler()
    floats_scaled = pd.DataFrame(scaler.fit_transform(
        floats), columns=floats.columns) * 0.2

    final_df = pd.concat([genre_df, floats_scaled, popularity_scaled,
                          subject_ohe, polar_ohe, key_ohe, mode_ohe], axis=1)
    final_df["id"] = df["id"].values

    return final_df


float_cols = tracks_df.dtypes[tracks_df.dtypes == "float64"].index.values
tracks_df.to_csv("tracks.csv", index=False)

features_df = create_features_dataset(tracks_df, float_cols=float_cols)
features_df.to_csv("features.csv")
