==> notebooks/04-recommendation.ipynb <==
# Sistema de recomendação

Neste notebook iremos fazer o sistema de recomendação

### Bibliotecas

- pandas

### Instalação e importação de dependências

pip install scikit-learn


import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity

Nesta etapa buscamos os `.csv` criados no notebook `03-process-data.ipynb` para fazermos as recomendações.

tracks_df = pd.read_csv("tracks.csv")
features_df = pd.read_csv("features.csv")


tracks_df.head()


features_df.head()


Além desses dois `.csv`, também precisamos de uma playlist de teste.

### Importar e pré processar playlist de teste

def drop_duplicates(df):
    df["artists_song"] = df.apply(lambda row: str(
        row["artist_name"]) + str(row["track_name"]), axis=1)
    return df.drop_duplicates("artists_song")


def select_columns(df):
    return df[["artist_name", "id", "track_name", "danceability", "energy", "key", "loudness", "mode",
               "speechiness", "acousticness", "instrumentalness", "liveness", "valence", "tempo", "artist_popularity", "genres", "track_popularity"]]


def process_genres(df):
    df["genres_list"] = df["genres"].apply(lambda x: str(x).split(" "))
    return df


def process_playlist(df):
    df = drop_duplicates(df)
    df = select_columns(df)
    df = process_genres(df)

    return df


playlist_df = pd.read_csv("playlist.csv")
playlist_df = process_playlist(playlist_df)
playlist_df.head()


def generate_playlist(features_df, playlist_df):
    # Encontra features na playlist
    features_in_playlist = features_df[
        features_df["id"].isin(playlist_df["id"].values)
    ]

    # Encontra features que não estão na playlist
    features_not_in_playlist = features_df[
        ~features_df["id"].isin(playlist_df["id"].values)
    ]
    features_in_playlist = features_in_playlist.drop(columns="id")

    return features_in_playlist.sum(axis=0), features_not_in_playlist


# Gera as features
features_in_playlist, features_not_in_playlist = generate_playlist(
    features_df, playlist_df
)


features_not_in_playlist.head()

features_in_playlist.head()

def get_recommendations(tracks_df, features_in_playlist, features_not_in_playlist):
    non_playlist_df = tracks_df[
        tracks_df["id"].isin(features_not_in_playlist["id"].values)
    ]

    non_playlist_df["sim"] = cosine_similarity(
        features_not_in_playlist.drop("id", axis=1).values,
        features_in_playlist.values.reshape(1, -1),
    )[:, 0]

    predict_playlist = non_playlist_df.sort_values("sim", ascending=False).head(50)

    return predict_playlist

recommended_playlist = get_recommendations(
    tracks_df, features_in_playlist, features_not_in_playlist
)
recommended_playlist.head()


recommended_playlist[["artist_name","track_name"]][:50]

recommended_playlist.to_csv("recommended_playlist.csv", index=False)
